refactor: Move useWebSiteHandler to hooks directory and update imports
- Created a new file for useWebSiteHandler in the hooks directory to improve code organization.
- Updated all relevant imports across the Discovery module to reference the new location of useWebSiteHandler.
- Removed the old utils directory reference to streamline the codebase.
